requirements.txt
nodriver>=0.38
google-genai>=1.0.0
pymupdf>=1.25.0
python-dotenv>=1.0.0
